backend-python/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic_settings
psycopg2-binary
alembic
pydantic[email]